+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Upgraded documentation.
 ### Removed
 - Onnx dependencies.
+- Python 3.9 support.
 
 ## [1.0.5] - 2024-12-13
 ### Added