+++ docs/CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.0] - 2025-01-29
+### Added
+- Support for traffic signals detection with a new YoloV11 model.
+- Checks on video and gps durations.
+### Changed
+- Upgraded ultralytics to 8.3.0.
+- Upgraded documentation.
+### Removed
+- Onnx dependencies.
+- Python 3.9 support.
+
 ## [1.0.5] - 2024-12-13
 ### Added
 - Add GPX format in gps sources.